pyproject.toml
[project]
name = "understanding-elastic-stack"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "dramatiq[redis]>=2.1.0",
    "fastapi>=0.135.3",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "sqlalchemy>=2.0.48",
    "uvicorn[standard]>=0.42.0",
]
